requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
passlib
onetimepass
bcrypt
pyotp
sqlalchemy
plotly
twilio
duckdb

